arreglo funcionalidad de subir imagen(libros)

+++ proyecto/src/main/java/com/Proyecto/repository/LibrosRepository.java
@@ -6,6 +6,6 @@ import org.springframework.stereotype.Repository;
 import com.Proyecto.model.Libro;
 
 @Repository
-public interface LibrosRepository extends JpaRepository<Libro, Long> {
+public interface LibrosRepository extends JpaRepository<Libro, Integer> {
 
 }
